unify style for language selectro
